requirements.txt
asgiref==3.7.2
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environs==10.3.0
marshmallow==3.20.2
packaging==23.2
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.4
sqlparse==0.4.4
tzdata==2023.4
